ad717x: add function to support to read data and status register together
Devices from the ad717x family and ad4111 support a feature where the
data register is output along with the status register to indentify the
channel from which the data came. This means that the SPI transaction is
longer by one byte at each data read.

Add functionality that takes into account the data length and the status
output option and calculates the SPI transacion length at every data read.

Signed-off-by: Andrei Drimbarean <Andrei.Drimbarean@analog.com>